refactor(core): Remove submit-workflow tool as it is dead code
